% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGSlidar.R
\name{buildPDALPipelineENTWINE}
\alias{buildPDALPipelineENTWINE}
\title{USGS Lidar Toolkit -- Create a PDAL pipeline to retrieve data from the ENTWINE collection}
\usage{
buildPDALPipelineENTWINE(
  inputPolygons,
  IDColumnLabel,
  URLColumnLabel = "url",
  pipelineOutputFileBaseName = "",
  pipelineOutputFolder = ".",
  pipelineTemplateFile = "",
  clipOutputFolder = ".",
  pipelineScript = "RUNME.bat",
  compress = TRUE,
  verbose = 0
)
}
\arguments{
\item{inputPolygons}{\code{Spatial*} or \code{sf} object containing a polygon(s) for
which you want to retrieve data. The polygons must include a field named \code{url}
that contains the URL for the ENTWINE data (ept.json) containing the polygon or the
name of the column containing the URL must be specified in \code{URLColumnLabel}.}

\item{IDColumnLabel}{Character: Name of the column in \code{inputPolygons} to be used to build
the output file name for the retrieved point data.}

\item{URLColumnLabel}{Character: Name of the column containing the URL for the ENTWINE data.
Default is "url". For Entwine data, this is typically the URL for the ett.json file.}

\item{pipelineOutputFileBaseName}{Character: base name for the new PDAL pipeline. This is
used along with the value in \code{IDColumnLabel} to create the actual file name. If not
specified, the pipeline file name will be created using the base name of the value
in the \code{url} field and the value in \code{IDColumnLabel}. \code{pipelineOutputFileBaseName}
should not contain any path information.}

\item{pipelineOutputFolder}{Character: full path name for the new PDAL pipeline(s). This should
be a fully qualified path. Pipelines will be saved in this folder. If not specified,
pipelines will be saved in the current working folder. This folder will be created
if it does not exist.}

\item{pipelineTemplateFile}{Character: full file specifier for the JSON file used
as a template for the output PDAL pipeline. See the description for the required items in
the template. A simple default template will be used if \code{pipelineTemplateFile}
is not provided.}

\item{clipOutputFolder}{Character: full path name for the point data file(s). This should
be a fully qualified path. Point files will be saved in this folder. If not specified,
point files will be saved in the current working folder. This folder will be created
if it does not exist. the file name for the point data file(s) is created using the
value in the \code{url} and \code{IDColumnLabel} fields. The file name extension will be
".las" if \code{compress = FALSE} or ".laz" if \code{compress = TRUE} or \code{compress}
is not specified.}

\item{pipelineScript}{Character: file name for the script to run the pipelines. For Windows
systems, this will be a batch file. No other operating systems are supported. The script
will be written in the folder defined by \code{pipelineOutputFolder}.}

\item{compress}{Logical: If TRUE (default), point data area stored in LAZ format; if FALSE,
point data are stored in LAS format.}

\item{verbose}{if > 0, output status information every \code{verbose} features
in \code{inputPolygons}.}
}
\value{
Integer: (invisible) number of pipeline files created.
}
\description{
Generate PDAL pipeline(s) to retrieve point data for a set of polygons. Given the input polygon(s),
create PDAL pipeline(s) to retrieve data for each polygon. Input data must be polygons and must have
a field that contains the URL for the ept.json file associated with the ENTWINE data set.
Input polygons can be either an \code{sp} or \code{sf} feature set.
}
\details{
The pipeline template must include at least the following items for the ept_reader: bounds
  and placeholder, and for the las_writer: filename. There is a template provided with the
  library that retrieves all data within the bounding box for a feature and stores is in compressed
  (LAZ) format. Users can supply their own template (in json format) but only the values
  mentioned for the ept_reader and las_writer tags will be modified.

  The pipeline template provided with the USGSlidar package is stored in the install folder
  for the package under the \code{extdata} folder. While you can directly modify this copy
  of the file, it would be better to copy the file to another location, make the desired
  changes and then specify your new template using the \code{pipelineTemplateFile} parameter.
}
\examples{
\dontrun{
buildPDALPipelineENTWINE(plot_polys)
}
}
