% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{clearUSGSProjectIndex}
\alias{clearUSGSProjectIndex}
\title{USGS Lidar Toolkit -- Clear the Stored Project Index File Name and
Boundary Cache}
\usage{
clearUSGSProjectIndex()
}
\value{
An (invisible) integer code, always 0.
}
\description{
Clear the name of the local project index file and remove the project
boundaries if they have been loaded. The loaded project boundaries
amount to a large amount of data. If you plan to save your workspace,
it is a good idea to call clearUSGSProjectIndex() to remove the
boundaries from memory. Otherwise expect large workspace files and
slow saves and loads.
}
\examples{
clearUSGSProjectIndex()
}
