% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGSlidar.R
\name{computeClipBufferForCONUS}
\alias{computeClipBufferForCONUS}
\title{USGS Lidar Toolkit -- Compute buffer size(s) to correct for distance
distortions associated with the web Mercator projection (EPSG:3857)}
\usage{
computeClipBufferForCONUS(
  desiredBuffer,
  centerPointLatitude = NULL,
  points = NULL
)
}
\arguments{
\item{desiredBuffer}{A single numeric value defining the desired buffer
size or a vector of numeric values with the desired buffer sizes. If a
vector of sizes is provided, it must be the same length as the vector
provided for \code{centerPointLatitude}.}

\item{centerPointLatitude}{A single numeric value or vector of values
containing the latitude of the sample points.}

\item{points}{A \code{SpatialPoints*} or \code{sf} object containing
the sample point locations. The object must contain a coordinate system
definition corresponding to the web Mercator projection. The points
will be projected into NAD83 lon-lat and the resulting latitude values
will be used to compute the buffer size(s).}
}
\value{
A single value that is the buffer width needed to produce the
  \code{desiredBuffer} or a vector of buffer widths needed to produce the
  \code{desiredBuffer} for each sample point.
}
\description{
The web Mercator projection does not maintain accurate distances. The
distortions increase as you move away from the equator so a correction
is needed to produce sample areas that are consistent in size for
a set of samples across a range latitudes.
}
\details{
\code{computeClipBufferForCONUS} uses a correction multiplier developed
for latitudes ranging from 23-49 degrees by comparing 1000m squares
defined in the web Mercator projection to their sized when reprojected
into UTM. While this is still not a perfect solution, it provides
a correction factor that gets you very close to the desired sample
size.
}
\examples{
\dontrun{
computeClipBufferForCONUS(1000, 41.5)
}
}
