% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{fetchFile}
\alias{fetchFile}
\title{USGS Lidar Toolkit -- Retrieve a file}
\usage{
fetchFile(url, destfile, method = "libcurl", mode = "wb", ...)
}
\arguments{
\item{url}{A character string with the URL for the file. If using local files,
the URL should start with \code{file://} and you should set \code{method="auto"}.}

\item{destfile}{A character string specifying the file name for the
downloaded file.}

\item{method}{Method used with \code{download.file()}. Refer to \code{download.file()}
documentation for possible methods.}

\item{mode}{A character string indicating the mode with which to write the
file. Refer to \code{download.file()} documentation for possible modes.}

\item{...}{Additional arguments passed to \code{download.file()}.}
}
\value{
An (invisible) integer code, 0 for success and non-zero for failure.
  This is the value returned from \code{download.file}.
}
\description{
This is a helper function used in the USGSlidar package to call
\code{utils::download.file()} to retrieve individual index files and data
tiles.
}
\examples{
\dontrun{
fetchFile("sample", "ftp://somewebsite.com/files/samplefile.bin")
}
}
