% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGSlidar.R
\name{fetchUSGSProjectIndex}
\alias{fetchUSGSProjectIndex}
\title{USGS Lidar Toolkit -- Download Lidar Project Index}
\usage{
fetchUSGSProjectIndex(
  destfile,
  url = "",
  type = "FESM",
  method = "libcurl",
  ...
)
}
\arguments{
\item{destfile}{A character string specifying the file name for the downloaded
index. If not provided or empty, the index is stored in the current
working directory using the filename in the url.}

\item{url}{A character string with the URL for the project index file. If an empty
string, the URL for the project index corresponding \code{type} is fetched.
When \code{type = "FESM"}, the URL for FESM_LPC_Proj.gpkg is used. When
\code{type = "WESM"}, the URL for WESM.gpkg is used. When \code{type = "ENTWINE"}
is used, the URL for the USGS Entwine data index maintained by Howard
Butler is used.}

\item{type}{A character string specifying the index to use. Valid values
are \code{"WESM"}, \code{"FESM"}, or \code{"ENTWINE"}.}

\item{method}{Method used with \code{download.file()}.  Refer to \code{download.file()}
documentation for possible methods.}

\item{...}{Additional arguments passed to \code{download.file()}.}
}
\value{
An (invisible) integer code, 0 for success and non-zero for failure.
  This is the value returned from \code{download.file}.
}
\description{
Retrieve the USGS lidar project index file. The project index file
contains project area boundaries for all USGS lidar data. Additional
attributes provide information about the acquisition such as the
starting and ending collection dates, pulse density, quality level,
work unit identifier and the work package identifier. The index is
stored in geopackage format and can be somewhat large. The index
is refreshed whenever new data is added to the USGS collection.
}
\examples{
\dontrun{
fetchUSGSProjectIndex("")
fetchUSGSProjectIndex("Project.gpkg")
fetchUSGSProjectIndex("Project.gpkg", method = "wininet")
}
}
