% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchFunctions.R
\name{fetchUSGSTiles}
\alias{fetchUSGSTiles}
\title{USGS Lidar Toolkit -- Download Lidar Tiles}
\usage{
fetchUSGSTiles(destfolder = "", urls, refresh = FALSE, method = "libcurl", ...)
}
\arguments{
\item{destfolder}{A character string with the name of the folder
where the downloaded tiles should be saved. If empty, tiles
are saved in the current working directory. The folder name
should not include a trailing slash. The folder will be created
if it does not exist.}

\item{urls}{A list of character strings with the URLs for tiles to
fetch.}

\item{refresh}{Logical value indicating files that already exist in
\code{destfolder} should be fetched again. The default behavior is
to skip existing files. For skipped files, the return value is set
to \code{0} indicating a successful fetch. The default behavior can
be problematic when you have \code{urls} that targets files in
different folders but have the same file name.}

\item{method}{Method used with download.file()}

\item{...}{Additional arguments passed to download.file()}
}
\value{
An (invisible) atomic vector of integer codes corresponding to the
  \code{urls}. The values are those returned from the \code{download.file}
  function calls for each of the \code{urls}, 0 for success and non-zero
  for failure. You can easily count the number of \code{urls} successfully
  fetched using \code{sum(value == 0, na.rm = TRUE)}.
}
\description{
Download one or more files (usually lidar data tiles) described
by a list of URLs and place them in a folder.
}
\examples{
\dontrun{
fetchUSGSTiles("", URLlist)
}
}
