% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryFunctions.R
\name{queryUSGSTileIndex}
\alias{queryUSGSTileIndex}
\title{USGS Lidar Toolkit -- Identify USGS Lidar Tiles Covering Area}
\usage{
queryUSGSTileIndex(
  x,
  y,
  buffer = 0,
  projectID = NULL,
  fieldname = "workunit_id",
  shape = "square",
  aoi = NULL,
  crs = "",
  index = "",
  segments = 60,
  return = "index",
  returnType = "sf",
  returncrs = "same",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Location easting for the center of the area-of-interest.}

\item{y}{Location northing for the center of the area-of-interest.}

\item{buffer}{Distance in meters added or subtracted to \code{(x,y)}
to create the area-of-interest. Can be 0 when used with \code{(x,y)} to
return items providing coverage for point location(s). Can be a
vector of values corresponding to the number of objects in \code{aoi}
allowing a different buffer size for each object when \code{aoi} is a
\code{Spatial*} or \code{sf} object. \code{buffer} can be negative with
polygon features to reduce the area considered. However, you may end up
with weird shapes if the distance is larger than the width of the polygon.
Data in \code{(x,y)} and \code{aoi} are projected into the web mercator
coordinate reference system before the buffer operation is done so the
units for \code{buffer} are always meters.}

\item{projectID}{Character string or list of character strings
containing the ID(s) for the lidar project(s). Typically obtained
by calling \code{queryUSGSProjectIndex()}. You must provide at least one
project identifier.}

\item{fieldname}{Character string containing the name of the field with
matches for the \code{projectID} to be used when querying the tile index.}

\item{shape}{Character string describing the shape of the sample area in
the case of point features or the shape applied to the buffer corners
when using polygon features. Valid values are \code{"square"} or \code{"circle"}.}

\item{aoi}{\code{Spatial*} or \code{sf} object containing a point or polygon
describing the area of interest. Can be points or polygons.}

\item{crs}{Valid \code{proj4string} string defining the coordinate
reference system of \code{(x,y)}. \code{crs} is required when using
\code{(x,y)}. \code{crs} is ignored when \code{aoi} is specified.
\code{crs} can also be an EPSG code (numeric).}

\item{index}{Index file for USGS lidar tiles If not provided, an index
previously specified by a call to \code{fetchUSGSProjectIndex} or
\code{setUSGSProjectIndex} will be used. If not provided and you have
not called \code{fetchUSGSTileIndex} or \code{setUSGSTileIndex}, the
function displays a warning and returns.}

\item{segments}{Number of segments to use when generating a circular
area of interest. When using a \code{SpatialPoint*} or \code{sf} object
with \code{shape = "circle"}, set \code{segments} to a rather large value (60
or higher) that is a multiple of 4. The \code{gBuffer} function from
\code{rgeos} and \code{st_buffer} function from \code{sf} are used to
build the sample areas and it accepts the number of segments in a quarter
circle so small values for \code{segments} may not produce good circles.
Values for \code{segments} that are not a multiple of 4 will not
produce circles with the correct number of segments.}

\item{return}{Character string specifying the kind of shapes to be
returned. Valid values are "\code{"aoi"} to return shapes for each
object in the \code{aoi} or the object defined by \code{(x,y)} and \code{buffer}
with additional attributes providing lidar project information and
\code{"index"} (default) to return tile shapes that provide coverage
for the features in the \code{"aoi"} or the area defined by \code{(x,y)}
and \code{buffer}. When tile shapes are returned there is no \code{"aoi"}
information included with the project shape attributes.}

\item{returnType}{Character string specifying the object type for the
returned polygon object when \code{(x,y)} is used to specify the
area-of-interest. Valid values are \code{"Spatial"} or \code{"sf"}.
\code{returnType} is ignored when \code{aoi} is specified.}

\item{returncrs}{Valid \code{proj4string} string defining the coordinate
reference system of the returned \code{SpatialPolygonsDataFrame} or
\code{sf} object. A value of \code{"same"} will project the return object
to the same coordinate reference system as \code{aoi} or to \code{crs} when
used with \code{(x,y)}.}

\item{verbose}{Boolean to enable printing of status messages. This is really
only useful for code debugging.}

\item{...}{Additional arguments passed to \code{download.file}}
}
\value{
A \code{SpatialPolygonsDataFrame} or \code{sf} object containing
  polygon(s) and attribute(s) for tiles covering the specified area.
  Attributes for the \code{aoi} are placed first in the \code{data.frame}
  followed by attributes for the lidar tile polygons..
}
\description{
Intersect a set of features (points or polygons) against a lidar
tile index to determine which tiles are needed to provide coverage for the
features. This involves both a spatial overlay operation and a database query.
The database query is used to limit the features used for the spatial
overlay. This dramatically increases the speed of the query when the
area of interest only involves a few lidar projects.
}
\details{
Query the tile index to find tiles that intersect the \code{buffer}ed
  \code{(x,y)} point(s) or \code{buffer}ed feature(s) provided in \code{aoi}.
}
\examples{
\dontrun{
queryUSGSTileIndex(-13540901, 5806426, 180, shape = "circle", crs = 3857)
  }
}
