% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGSlidar.R
\name{setUSGSProjectIndex}
\alias{setUSGSProjectIndex}
\title{USGS Lidar Toolkit -- Set Filename for the Lidar Project Index}
\usage{
setUSGSProjectIndex(index)
}
\arguments{
\item{index}{Character string with the local file name for the project index.}
}
\value{
An (invisible) integer code, 0 for success and non-zero for failure.
}
\description{
Set the filename and path for the local copy of the USGS lidar project
index file. This is used when you have already used fetchUSGSProjectindex
to download a project index. While the project index is frequently
updated by the USGS, the index is large and it may not be necessary
to download a new copy every time you want to download lidar data.
}
\examples{
\dontrun{
setUSGSProjectIndex("g:/working_data/ProjectIndex.gpkg")
}
}
