% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPCQueryFunctions.R
\name{signMPCURL}
\alias{signMPCURL}
\title{USGS Lidar Toolkit -- Sign lidar tile URLs using MPC access token}
\usage{
signMPCURL(URLs = NULL)
}
\arguments{
\item{URLs}{character list containing unsigned URL(s) for point tiles.}
}
\value{
A character list of signed URLs that can be used with \code{fetchUSGSTiles}
  or any other function that retrieves files using https protocol.
}
\description{
Query Microsoft Planetary Computer (MPC) to obtain an access token and use it to sign
URLs for lidar point tiles. Function does not verify that the \code{URLs} point to
assets on MPC. It does, however, verify that \code{URLs} point to assets in
MPC blob storage. Only assets in MPC blob storage need access tokens. If \code{URLs}
do not point to assets in MPC blob storage, \code{URLs} are returned unchanged.
}
\details{
Existing token are removed from \code{URLs} prior to signing.

Signed URLs are only valid for the current R session and cannot be stored
for later use without resigning the URL.
}
\examples{
\dontrun{
signMPCURL(URLs)
  }
}
