% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{USGSProjectIndexIsCached}
\alias{USGSProjectIndexIsCached}
\title{USGS Lidar Toolkit -- Check For Project Index File Name and
  Boundary Cache}
\usage{
USGSProjectIndexIsCached()
}
\value{
An (invisible) integer code, always 0.
}
\description{
Checks to see if there is a cached set of project polygons. Project
boundaries (index file name and polygons) are saved the first time
\code{queryUSGSProjectIndex} is called to speed up subsequent queries.
\code{USGSProjectIndexIsCached} lets you check to see if there is a
set of cached index polygons so you can better manage the \code{index}
parameter in calls to \code{queryUSGSProjectIndex}.
}
\examples{
USGSProjectIndexIsCached()
}
